Moving mongo pruner into mongo package
