+++ src/app/beer_garden/db/mongo/pruner.py
